Progress on stateManager
